Move FsPath and AnyPath to core.fs.path
