+++ parsec/core/fs/__init__.py
@@ -42,13 +42,16 @@ from parsec.core.fs.exceptions import (
     FSDeviceNotFoundError,
     FSInvalidTrustchainEror,
 )
+from parsec.core.fs.path import FsPath, AnyPath
 from parsec.core.fs.workspacefs import WorkspaceFS, WorkspaceFSTimestamped
-
 
 __all__ = (
     "UserFS",
     "WorkspaceFS",
     "WorkspaceFSTimestamped",
+    # Path
+    "FsPath",
+    "AnyPath",
     # Generic error classes
     "FSError",
     "FSOperationError",
