+++ parsec/core/fs/exceptions.py
@@ -10,16 +10,19 @@ Define all the FSError classes, using the following hierarchy:
         +-- FSRemoteOperationError
 
 """
+from __future__ import annotations
 
 import os
 import errno
 import io
 
-from parsec.core.types import EntryID, ChunkID, AnyPath
+from parsec.core.types import EntryID, ChunkID
 from parsec.core.fs.utils import ntstatus
 
-from typing import Optional, Union
+from typing import Optional, Union, TYPE_CHECKING
 
+if TYPE_CHECKING:
+    from parsec.core.fs.path import AnyPath
 
 # Base classes for all file system errors
