+++ parsec/core/types/__init__.py
@@ -4,8 +4,6 @@ from typing import Union, NewType
 
 from parsec.api.data import EntryID, EntryIDField, EntryName, EntryNameField
 
-from parsec.core.types.base import FsPath, AnyPath
-
 from parsec.core.types.backend_address import (
     BackendAddr,
     BackendOrganizationAddr,
@@ -53,9 +51,6 @@ __all__ = (
     "LocalFolderishManifests",
     "RemoteFolderishManifests",
     "LocalNonRootManifests",
-    # Base
-    "FsPath",
-    "AnyPath",
     # Entry
     "EntryID",
     "EntryIDField",
