+++ tests/core/fs/test_path.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from parsec.core.types import FsPath
+from parsec.core.fs import FsPath
 
 
 @pytest.mark.parametrize(
