+++ tests/core/fs/workspacefs/test_confinement.py
@@ -3,7 +3,7 @@
 import re
 import pytest
 
-from parsec.core.types import FsPath
+from parsec.core.fs import FsPath
 from tests.common import create_shared_workspace
