+++ tests/core/fs/workspacefs_timestamped/test_entry_transactions_timestamped.py
@@ -3,7 +3,7 @@
 import pytest
 from pendulum import datetime
 
-from parsec.core.types import FsPath
+from parsec.core.fs import FsPath
 from parsec.core.fs.exceptions import FSLocalMissError
