+++ tests/core/fs/workspacefs_timestamped/test_file_transactions_timestamped.py
@@ -2,8 +2,7 @@
 
 import pytest
 
-from parsec.core.types import FsPath
-from parsec.core.fs import FSError, FSInvalidFileDescriptor
+from parsec.core.fs import FsPath, FSError, FSInvalidFileDescriptor
 
 
 @pytest.mark.trio
