+++ tests/core/fs/workspacefs_timestamped/test_workspacefs.py
@@ -3,7 +3,7 @@
 import pytest
 from pendulum import datetime, now
 
-from parsec.core.types import FsPath
+from parsec.core.fs import FsPath
 from parsec.core.fs import FSError
