+++ tests/core/fs/workspacefs_timestamped/test_workspacefs_timestamped.py
@@ -3,7 +3,7 @@
 import pytest
 from unittest.mock import ANY
 
-from parsec.core.types import FsPath
+from parsec.core.fs import FsPath
 from parsec.core.fs.exceptions import FSWorkspaceTimestampedTooEarly
