+++ tests/scripts/debug_workspace.py
@@ -10,7 +10,7 @@ from humanize import naturalsize
 
 from parsec.logging import configure_logging
 from parsec.core.logged_core import logged_core_factory
-from parsec.core.types import FsPath
+from parsec.core.fs import FsPath
 from parsec.core.config import get_default_config_dir, load_config
 from parsec.core.local_device import list_available_devices, load_device_with_password
 from parsec.test_utils import (
